BackEnd/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient 
validate_email
py3DNS
django-filter
pyrebase4
jsons
django-cleanup